requirements.txt
flask
flask-cors
python-dotenv
httpx
google-generativeai
chromadb